Added option to run the transfer to PIO in a critical section
This eliminates the glitching described here: https://github.com/blaz-r/pi_pico_neopixel/issues/12

Depending on how long the strip is, this could be a pretty long period with interrupts disabled.  But I suspect it will really be the most ciritical task in a lot of applications and they will benefit from disabling the interrupt.